Pydantic_playground/requirements.txt
# requirements.txt
python-dotenv
requests
aiohttp
beautifulsoup4
lxml
